requirements.txt
# ClawMark Generator Dependencies
# Web templating and configuration
jinja2>=3.1.0
pyyaml>=6.0

# Optional: For advanced features
pillow>=9.0.0  # For image processing if needed
requests>=2.28.0  # For fetching remote assets

# Development dependencies (optional)
# black>=22.0.0  # Code formatting
# flake8>=5.0.0  # Linting